requirements.txt
flask==3.0.3
werkzeug==3.0.3
jinja2==3.1.4
flask-restful==0.3.10
paho.mqtt==2.1.0